fix usage of gevent.joinall, increase timeouts
We did run into errors on the production server. Some of the calls to get events
did timeout, and we ran into errors because the code didn't handle that case.

gevent.joinall returns the list of finished greenlets, but most of the code
tried to collect results from all of the greenlets.