add ethindex_db
ethindex_db provides access  to the ethindex postgres database. The main class
tries to be a drop-in replacement for the parts of the CurrencyNetworkProxy
class that's being used from relay.api in order to get events from the
blockchain.

This isn't finished yet, only get_events is implemented at this point.