travis: use build stages
We use build stages, so that we have a chance to hide our DOCKER_PASSWORD.

The Dockerfile itself now also uses a staged build and copies the virtualenv to
/opt/relay. The base image itself hasn't changed, i.e. I haven't upgraded to
ubuntu 18.04.

The build is pretty much consistent now with what we're doing in the watch and
py-eth-index repositories.

This will also trigger the deploy stage for tags.