Query the nonce from pending transactions
This enables a more up to date nonce information. It will use the pending block to determine the next nonce or on parity it will use the nextNonce function that also includes the pending transactions. At the moment it will always default to parity, as we only use parity anyways. This also requires that the parity api is enabled (--jsonrpc-apis=all or --jsonrpc-apis=parity)