don't use the same psql connection concurrently
We're not allowed to use the same connection concurrently from multiple
greenlets.

This patch adds a concurrency_utils.synchronized decorator, that synchronizes
method access. It's being used to synchronize access to the OrderBookDB object.

This introduces a new dependency on the BSD licensed wrapt library. Please visit
https://github.com/GrahamDumpleton/wrapt/blob/develop/blog/07-the-missing-synchronized-decorator.md
for a discussion of the synchronized decorator and
https://github.com/GrahamDumpleton/wrapt/tree/develop/blog for a whole series on
the wrapt library and decorators.

see https://github.com/trustlines-network/relay/issues/197 and
https://github.com/trustlines-network/relay/issues/142