use ethindex even if ETHINDEX environment variable is not set
Running without the ethindex does not work anymore in the web3 upgrade branch.
We would either have to fix it or remove it. I would choose the latter as I
don't see any good reason to keep it.

In order to use the non-ethindex version, ETHINDEX=0 now has to be set
explicitly.

I failed to set ETHINDEX=1 multiple times in the past, so this will help me a
bit.