add a workaround for running the end2end tests
the end2end tests don't run when we we relay multiple transactions at the same
time.

It looks like parity just does not create a new block for the second transaction
it sees though we are using the instant seal engine.

When we make the relay_tx method synchronized, transactions will not end up at
parity at the same time. This makes it possible to run the end2end tests.

Somehow this only became an issue after the upgrade to web3 4.x

See https://github.com/paritytech/parity-ethereum/issues/4494