use dump_result_with_schema decorator
Instead of calling dump on some half finished data structure, dumping it and
extending it with additional fields, we now build the whole result and dump it
once at the end of the function.

TrustlineSchema has been deleted as it's unused now. It was equal to the
AccountSummarySchema. Not sure what was intended here.

get_account_sum should probably be split into the two methods that are
implemented inline. I fail to see a reason why it should be one method.