add an implementation of dijkstra path finding
This adds a least_cost_path function, which implements dijkstra's
multi-source/multi-target path finding with a configurable cost function.

This introduce a CostAccumulator abstract class, which is being used to compute
the total cost for a path when hopping over the next edge. This is pretty much
exactly what we need when 'transferring values' inside the trustlines graph.

We should be able to use this for both the 'sender pays' and the 'receiver pays'
payment methods.

This is based on ubergraph's implementation in clojure (see
https://github.com/Engelberg/ubergraph).

We don't have a way to filter nodes or edges at the moment.

Currently there are not tests, though I do have some code that still needs some
tidying up.

My plan for integrating this (i.e. replacing dijkstra_weighted.find_path) still
includes fixing the find_path function first and adapt tests to the fixed
function (in case some of the tests do break).