implement find_path by calling into least_cost_path
We sort by the fees first and then by number of hops. The original find_path
couldn't really decide on what to sort on [1].

Sorting by fees first will give the network a chance to have better rebalancing
by reducing the global imbalance. The downside of course is that it may use more
hops for payments, which results in higher gas costs.

This also fixes the issue mentioned in [1].

[1] https://github.com/trustlines-network/relay/issues/219