add Graph.close_trustline_path_triangulation function
This computes a payment path to be used for closing a trustline.

We now have the SenderPaysCostAccumulatorSnapshot
and ReceiverPaysCostAccumulatorSnapshot CostAccumulators, which can be used to
compute the cost for given timestamp. The former is similar to the
dijkstra_weighted.FeesFirstSenderPaysCostAccumulator, though it allows to set a
constant timestamp and it has the fee computation 'baked in'.