handle compute_cost_for_path being called with a forbidden path
compute_cost_for_path may not be able to compute the cost for a given path
because it may exceed the max_hops or max_fee boundaries.

In that case we now raise a nx.NetworkXNoPath exception. This might seem
arguable, but it's consistent with what _least_cost_path_helper does.

We also rely on this behaviour in Graph.close_trustline_path_triangulation now
and return an empty PaymentPath in that case.