add a way to generate and use testdata from deployed contracts
This commit introduces a way to generate test data from deployed contracts and
use that test data to generate tests for our python code.

At the moment we use this only to test our calculate_fees and
calculate_fees_reverse functions, which are also part of this commit.

The gen_testdata.py script generates pseudo-random input data, calls the test
contract and writes a file with the input data and the expected results.

A hook in conftest.py generates test cases for this input data, if the test
function takes certain arguments.

This whole thing is currently in its infancy and one could argue that it's
overkill to test those two functions. So, please bear with me.