fix ReceiverPaysCostAccumulatorSnapshot
When the receiver pays, the fees are only paid if there is a next hop. They are
paid because the mediator forwards less money.

This means we need to add an additional field to our cost tuple (the next fee to
be paid out).

This fixes all but one the unit tests. Integration tests need fixing to (I
guess).