Refactor max capacity tests
Refactor max capacity tests to have not so much duplicated code. It does
not remove all the duplication, as I decided it is not worth it at the
moment. If this causes trouble, it should be done when it does so.

Contense all the tests that have arbitrary data and test that the path
is the maximum path into one test.

Provide a fixtures to easily configure the balances and the whole graph