Change documentation and example to use tlbc
Change the documentation and the docker-compose example
of the relay server setup to use the trustline blockchain. For the
documentation on how to run the blockchain node itself, we refer to the
documentation in the github repo. The example setup shows how to run the
relay server to work with our predeployed contracts.

Closes #241