Remove time.time() from graph
Remove the usage of time.time() completly from graph to make it not
depending in the time. Also remove time.time() from event creation, the
timestamp has now to be provided.
Changed the default parameter of timestamp from None to 0, so that you
can still not provide it if the timestamp is not relevant. To protect
from accidentially not providing the timestamp when it is relevant, I
added a assertion in the relevant calc interest calculation.