add a schema for meta transactions
This adds a marshmallow schema for meta transactions. I've had to introduce a
new HexEncodedBytes field, because the relay.api.fields.HexBytes field does not
roundtrip correctly.

See the meta transaction format issue for a description of meta transactions:
https://github.com/trustlines-network/project/issues/361

This requires trustlines-contracts version 0.6.0. trustlines-contracts-deploy
became a real dependency with this commit now.