force black's target version to be python 3.6
We have some team members that reformat the files with a different line length
for working with them and formatting them with the default options before
committing.

This may introduce some unwanted changes, because black tries to automatically
determine the target python version and the determined version for the
reformatted file may differ. The reason seems to be that trailing whitespaces
get removed when using longer lines.

Also see black's description of the --py36 option:
,----
|   --py36            Allow using Python 3.6-only syntax on all
|                     input files.  This will put trailing commas
|                     in function signatures and calls also after
|                     *args and **kwargs. Deprecated; use
|                     --target-version instead. [default: per-file
|                     auto-detection]
`----