Fix chain integration test for failing deployment
The deployment of the new currency networks (by
trustlines-contracts-bin@1.0.0) with eth-tester fails due to too high
gas in relation to the gas limit. Even if the deployment of a currency
network contract should cost ~5.3E6 and the block limit is set to 6E6,
the deployment fails. This seems to be caused by a wrong gas estimation
and rounding up. There it has been decided to increase the block gas
limit to the max value of the tlbc chain with 8E6.

This commit does also include a fix for the tests caused by a rebase to
the restructuring on master.