First passing test for getting list of interests as written in contracts
The test was taken and slightly adapted from the contracts repository
The interests rates there were gotten hard coded so this version of the
function to get the list of paid interests has hard coded interest rate.
We still need to get the interests rate at given balance update
timestamp.