Make the relay server sign transactions
To make the relay server sign transactions locally, you must configure
the account the be used via a snippet like the following in the TOML
configuration file:

,----
| [relay.account]
| keystore_path = "path/to/keystore.json"
| keystore_password_path = "path/to/passwd.txt"
`----

We use middleware that automatically signs transactions.

This requires the latest contract-deploy-tools.

This is an experimental feature. There might be race conditions with
the nonce handling, because previously we relied on parity to handle
the nonce.