Remove data content of push notification
To be sure that the information shown in the app comes from the
connected relay server and not from push notifications, we remove the
data content. This might be added later again, to allow to at least send
the user to the right page in the app.
Also messages should only be sent out based on the underlying event and
does not depend on the message text.