Add config to enable/disable relay features
Add config parameters to turn on/off certain features.
Currently this is mainly used to enable/disable the api endpoints, but
can be later extended to also internally shut down the services.
To not break the current behaviour all features are enabled per default
as before. This will be changed.

Add a proper config validation via marshmallow. To not break it, the old
config is still accepted, but deprecated, and will be removed in the
future.