Make /networks endpoint faster
We were always fetching the users from the chain node when asking for a
network. Make the endpoint faster by using the internal graph representation. This
might be a little outdated, but as we are subscriped to the events it
should only be several seconds old.
We can in the future decide to remove the call to users completly, but i
did not want to add a breaking change now.

Also renamed some methods to make it clearer that they are slow.