Revert custom event identifier feature
During the praxis tests we discovered that the pending events where not
received by the mobile app anymore. During the inspection it turned out
that the event identifier calculation fails for pending events. The log
index is `None` and lead to a failure. Unfortunately this did not occur with the
end-to-end tests.
Since the custom event identifier was never finally implemented and was
still under discussion for pending events, the feature has been
postponed. The thought was that the additional response field does not
hurt since it is not used. But because it lead to issues in the staging
setup, we need to revert it here.

This reverts the commits:
  - a259d02115059b7e942b5faca56089aa2c732850
  - 5d8b99aaa68a21921449dd00cec3a8a1899b4958