Add documentation how to get contracts.json file
By user feedback it has turned out that it is kinda hard to get the
file with the compiled contracts. Therefore the documentation has been
extended to provide instructions how to download the file on different
platforms. It is bundled within multiple packages, so the user can
select the approach he prefers.

Related to: #439