Improve performance of user events endpoint
The user events endpoint was to slow, as it did a request per every
currency network seperatly. This was done because we were to lazy before
to do it properly. This commit improves the situtation by using only one
db call for a user event endpoint call. However, the solution is quick
and dirty and we should refactor it in the future.