Fix send str again over websocket
Because of a change in tinytpc we were sending bytes over the websocket,
which the other end could not handle properly. This was not an issue in
nodejs but in other environments, thats why we could not see this in the
e2e tests. Fixed by decoding it back to a string before sending it.