Send data messages instead of push notifications
With this commit the relay server will send data messages instead of push notifications. The data message will contain blockHash, blockNumber, transactionHash and eventType for blockchain events. And eventType and message props for non-blockchain events.

This way the receiving client can decide how to display and we are not disclosing personal information in the message (for blockchain events).

For the non-blockchain events we relay the sent message as we currently don’t store them on the relay server and there won’t be any way for clients to query the server to get them.