Add android high priorit and aps content_available
If we are sending a message that is data-only and the device is in the background or quit, both Android & iOS treat the message as low priority and will ignore it (i.e. no event will be sent). To go around this we set priority to high (Android) and content-available to true (iOS) properties on the payload.