Add coverage to requirements instead of dev-requirements
It is required to run the relay since we report coverage from the
relay running for the end2end tests.
The other choice would be to only import coverage when ran with
`--coverage` and leave coverage in dev-requirements but that would
mean that we would need to build a relay image with the dev-requirements
installed.